Functions

Splinart.circle — Method

circle(center, radius; npoints)

Discretization of a circle.

Parameters

center : 2d coordinates of the center.

radius : Radius of the circle.

npoints : Number of discretization points (the default value is 50).

Returns

The theta angle.

The 2d coordinates of the circle.

Splinart.draw_pixel! — Method

draw_pixel!(img, xs, ys)

Add pixels on the image.

Parameters

img : The image where we add pixels.

xs : The x coordinate of the pixels to add.

ys : The y coordinate of the pixels to add.

Splinart.line — Method

line(b, e; ypos, npoints)

Discretization of a horizontal line.

Parameters

begin : The left point of the line.

end : The right point of the line.

ypos : The position of the y coordinate (the default value is 0.5).

npoints : Number of discretization points (the default value is 50).

Returns

The 2d coordinates of the line.

Splinart.save_img — Method

save_img(img, path, filename)

Save the image in a png file.

Parameters

img : The image to save.

path : The save directory.

filename : The file name with the png extension.

Splinart.show_img — Method

show_img(img)

Plot the image using matplotlib.

Parameters

img : The image to save.

Splinart.spline — Method

spline(xs, ys)

Return the second derivative of a cubic spline.

Parameters

xs : The x coordinate of the cubic spline.

ys : The y coordinate of the cubic spline.

Returns

The second derivative of the cubic spline.

Splinart.splint! — Method

splint!(y, xs, ys, y2s, x)

Evaluate a sample on a cubic pline.

Parameters

xs : The x coordinates of the cubic spline.

ys : The y coordinates of the cubic spline.

y2s : The second derivative of the cubic spline.

x : The sample where to evaluation the cubic spline.

y : The y coordinates of the sample.

See Also

Splinart.spline

Splinart.update_img! — Method

update_img!(
    img,
    path,
    xs_func,
    x;
    nrep,
    scale_value,
    periodic
)

Update the image using a cubic spline on a shape.

Parameters

img : The output image.

path : The y coordinate of the cubic spline if x is not None, the coordinates of the cubic spline if x is None.

x : The x coordinates of the cubic spline if given. (the default value is None)

xs_func : The function that return the x coordinate of the sampling points where to compute the y coordinates given the spline equation.

nrep : Number of iteration (default is 300).

periodic : Define if the first and last points of the path must be equal (default is true).

scale_value : Rescale the random radius (default value is 0.00001).

See Also

update_path!

Splinart.update_path! — Method

update_path!(path; scale_value, periodic)

Update the path of the spline.

We move each point of the path by a random vector defined inside a circle where

the center is the point of the path

the radius is a random number between [-1, 1]

Parameters

path : The y coordinate of the cubic spline.

periodic : If true, the first and the last points of the path are the same (the default value is false).

scale_value : Rescale the random radius (default value is 0.00001).